Changed hourly forecast temp from double to int. Also, fixed 7 day forecast listview small icon issue

+++ app/src/main/java/com/georgepazdral/StormyPlus/weather/Hour.java
@@ -34,7 +34,7 @@ public class Hour implements Parcelable {
         mSummary = summary;
     }
 
-    public double getTemperature() {
+    public int getTemperature() {
         return (int) Math.round(mTemperature);
     }
